Crate faktory_async

This crate is an async implementation of the Faktory Protocol.

NB: This crate does not implement application level logic for producer/consumer operation. It is purely an asynchronous implementation of the Faktory Protocol itself.

Re-exports

pub use faktory_command::FaktoryCommand;

pub use frame::RequestFrame;

pub use frame::ResponseFrame;

pub use resp::ServerResponse;

Modules

faktory_command

frame

resp

Structs

ClientManager

The ClientManager allows us to fully utilize async for workers. Each client is a single connection in Faktory parlance. Incidentally, Faktory enterprise is limited on a per connection basis. Thus, if we can limit the number of connections, we can maximize efficiency in terms of Faktory, and wrt I/O bound jobs. The basic idea is that we only use a single client, and the ClientManager serves as a non blocking, multi-task enabled interface to that client. Instead of directly accessing a FaktoryClient, tasks instead send commands (messages) to a ClientManager over a channel, and the ClientManager dispatches them to the underlying FaktoryClient.

ClientOptions

Failure

FaktoryClient

This is the underlying async client for the Faktory Protocol. It is the lowest level implementation, and higher level libraries should most likely use ClientManager.

FaktoryCodec

A simple Decoder and Encoder implementation that splits up data into lines.

FaktoryJob

A Faktory job.

Enums

Command

Picket Line commands. This is somewhat higher level than crate::FaktoryCommand. Essentially, here we are interested with commands from the Picket Line API, as opposed to Faktory API.

JobFailedError

Type Definitions

ClientManagerReceiver

ClientManagerSender