Struct faktory_async::FaktoryJob

pub struct FaktoryJob {
    pub queue: String,
    pub kind: String,
    pub created_at: Option<DateTime<Utc>>,
    pub enqueued_at: Option<DateTime<Utc>>,
    pub at: Option<DateTime<Utc>>,
    pub reserve_for: Option<usize>,
    pub retry: Option<isize>,
    pub priority: Option<u8>,
    pub backtrace: Option<usize>,
    pub failure: Option<Failure>,
    pub custom: HashMap<String, Value>,
    // some fields omitted
}

A Faktory job.

See also the Faktory wiki.

Fields
queue: String

The queue this job belongs to. Usually default.

kind: String

The job’s type. Called kind because type is reserved.

created_at: Option<DateTime<Utc>>

When this job was created.

enqueued_at: Option<DateTime<Utc>>

When this job was supplied to the Faktory server.

at: Option<DateTime<Utc>>

When this job is scheduled for.

Defaults to immediately.

reserve_for: Option<usize>

How long to allow this job to run for.

Defaults to 600 seconds.

retry: Option<isize>

Number of times to retry this job.

Defaults to 25.

priority: Option<u8>

The priority of this job from 1-9 (9 is highest).

Pushing a job with priority 9 will effectively put it at the front of the queue. Defaults to 5.

backtrace: Option<usize>

Number of lines of backtrace to keep if this job fails.

Defaults to 0.

failure: Option<Failure>

Data about this job’s most recent failure.

This field is read-only.

custom: HashMap<String, Value>

Extra context to include with the job.

Faktory workers can have plugins and middleware which need to store additional context with the job payload. Faktory supports a custom hash to store arbitrary key/values in the JSON. This can be extremely helpful for cross-cutting concerns which should propagate between systems, e.g. locale for user-specific text translations, request_id for tracing execution across a complex distributed system, etc.

Implementations

impl FaktoryJob

pub fn new<S, A>(kind: S, args: Vec<A>) -> Self where
S: Into<String>,
A: Into<Value>,

Create a new job of type kind, with the given arguments.

pub fn on_queue<S: Into<String>>(self, queue: S) -> Self

Place this job on the given queue.

If this method is not called (or self.queue set otherwise), the queue will be set to “default”.

pub fn id(&self) -> &str

This job’s id.

pub fn kind(&self) -> &str

This job’s type.

pub fn args(&self) -> &[Value]

The arguments provided for this job.